Fixed issue with launching ucp site.
Renamed files which describe classes. __autoload function requires name
in format :. $classname . ".class.php"
