+++ include/ACL.class.php
@@ -2,20 +2,18 @@
 
 class ACL
 {
+  public function get()
+  {
+    global $sessionId, $response;
+    $masterTable= new table('master');
+    $master=$masterTable->loadById( $sessionId );
+    $userGroupId=$master['group'];
 
-    public function get()
-    {
-        global $sessionId, $response;
-        $masterTable = new table('master');
-        $master = $masterTable->loadById($sessionId);
-        $userGroupId = $master['group'];
-
-        $groupTable = new table('group');
-        $group = $groupTable->loadById($userGroupId);
-        $userAcl = json_decode($group['acl'], true);
-        $response['acl'] = $userAcl;
-    }
-
+    $groupTable= new table('group');
+    $group=$groupTable->loadById( $userGroupId );
+    $userAcl=json_decode($group['acl'], true);
+    $response['acl'] = $userAcl;
+  }
 }
 
 ?>
