+++ include/updaterouter.php
@@ -1,6 +1,7 @@
 <?php
 
 include_once "/core.php";
+
 $routerId = $argv[1];
 $mode = $argv[2];
 if (isset($argv[3]))
