+++ ucp/index.php
@@ -2,7 +2,6 @@
 
 include_once '../include/ucp.php';
 
-
 $mainPageText = configgetvalue('system', 'ucp', NULL, 'mainPageText');
 $mainPageText = preg_replace('%\n%', '<br>', $mainPageText);
 
@@ -13,4 +12,5 @@ $tpl = array(
 $fenom->display($theme->getTemplateLocation('header.tpl'), $headerData);
 $fenom->display($theme->getTemplateLocation('index.tpl'), $tpl);
 $fenom->display($theme->getTemplateLocation('footer.tpl'));
+
 ?>
